Move task handlers under the controllers directory
