+++ concrete/controllers/permissions/categories/task_handlers/board.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Concrete\Core\Permission\Category\TaskHandler;
+namespace Concrete\Controller\Permissions\Categories\TaskHandlers;
 
 use Concrete\Core\Controller\Controller;
 use Concrete\Core\Entity\Board\Board as BoardEntity;
