+++ concrete/controllers/permissions/categories/task_handlers/conversation.php
@@ -1,9 +1,9 @@
 <?php
 
-namespace Concrete\Core\Permission\Category\TaskHandler;
+namespace Concrete\Controller\Permissions\Categories\TaskHandlers;
 
 use Concrete\Core\Controller\Controller;
-use Concrete\Core\Entity\Calendar\Calendar as CalendarEntity;
+use Concrete\Core\Conversation\Conversation as ConcreteConversation;
 use Concrete\Core\Error\UserMessageException;
 use Concrete\Core\Http\ResponseFactoryInterface;
 use Concrete\Core\Permission\Access\Access;
@@ -12,13 +12,11 @@ use Concrete\Core\Permission\Category\TaskHandlerInterface;
 use Concrete\Core\Permission\Checker;
 use Concrete\Core\Permission\Duration;
 use Concrete\Core\Permission\Key\Key;
-use Concrete\Core\Workflow\Workflow;
-use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Component\HttpFoundation\Response;
 
 defined('C5_EXECUTE') or die('Access Denied.');
 
-class Calendar extends Controller implements TaskHandlerInterface
+class Conversation extends Controller implements TaskHandlerInterface
 {
     /**
      * {@inheritdoc}
@@ -27,37 +25,34 @@ class Calendar extends Controller implements TaskHandlerInterface
      */
     public function handle(string $task, array $options): ?Response
     {
-        $calendar = $this->getCalendar($options);
-        if ($calendar === null) {
-            throw new UserMessageException(t('Calendar not received'));
-        }
+        $conversation = $this->getConversation($options);
         $method = lcfirst(camelcase($task));
         if (!method_exists($this, $method)) {
             throw new UserMessageException(t('Unknown permission task: %s', $task));
         }
 
-        return $this->{$method}($calendar, $options);
+        return $this->{$method}($conversation, $options);
     }
 
-    protected function getCalendar(array $options): ?CalendarEntity
+    /**
+     * @throws \Concrete\Core\Error\UserMessageException
+     */
+    protected function getConversation(array $options): ?ConcreteConversation
     {
-        $calendarID = $options['caID'] ?? null;
-        $calendar = $calendarID ? $this->app->make(EntityManagerInterface::class)->find(CalendarEntity::class, $calendarID) : null;
-        if ($calendar === null) {
-            return null;
-        }
-        $cp = new Checker($calendar);
-        if (!$cp->canEditCalendarPermissions()) {
+        $conversationID = (int) ($options['cnvID'] ?? 0);
+        $conversation = $conversationID > 0 ? ConcreteConversation::getByID($conversationID) : null;
+        $cpn = new Checker($conversation);
+        if (!$cpn->canEditConversationPermissions()) {
             throw new UserMessageException(t('Access Denied.'));
         }
 
-        return $calendar;
+        return $conversation;
     }
 
-    protected function addAccessEntity(CalendarEntity $calendar, array $options): ?Response
+    protected function addAccessEntity(?ConcreteConversation $conversation, array $options): ?Response
     {
         $pk = Key::getByID($options['pkID']);
-        $pk->setPermissionObject($calendar);
+        $pk->setPermissionObject($conversation);
         $pa = Access::getByID($options['paID'], $pk);
         $pe = Entity::getByID($options['peID']);
         $pd = empty($options['pdID']) ? null : Duration::getByID($options['pdID']);
@@ -66,10 +61,10 @@ class Calendar extends Controller implements TaskHandlerInterface
         return $this->app->make(ResponseFactoryInterface::class)->json(true);
     }
 
-    protected function removeAccessEntity(CalendarEntity $calendar, array $options): ?Response
+    protected function removeAccessEntity(?ConcreteConversation $conversation, array $options): ?Response
     {
         $pk = Key::getByID($options['pkID']);
-        $pk->setPermissionObject($calendar);
+        $pk->setPermissionObject($conversation);
         $pa = Access::getByID($options['paID'], $pk);
         $pe = Entity::getByID($options['peID']);
         $pa->removeListItem($pe);
@@ -77,34 +72,42 @@ class Calendar extends Controller implements TaskHandlerInterface
         return $this->app->make(ResponseFactoryInterface::class)->json(true);
     }
 
-    protected function savePermission(CalendarEntity $calendar, array $options): ?Response
+    protected function savePermission(?ConcreteConversation $conversation, array $options): ?Response
     {
         $pk = Key::getByID($options['pkID']);
-        $pk->setPermissionObject($calendar);
+        $pk->setPermissionObject($conversation);
         $pa = Access::getByID($options['paID'], $pk);
         $pa->save($options);
-        $pa->clearWorkflows();
-        $wfIDs = $options['wfID'] ?? null;
-        if (is_array($wfIDs)) {
-            foreach ($wfIDs as $wfID) {
-                $wf = Workflow::getByID($wfID);
-                if ($wf !== null) {
-                    $pa->attachWorkflow($wf);
-                }
-            }
-        }
 
         return $this->app->make(ResponseFactoryInterface::class)->json(true);
     }
 
-    protected function displayAccessCell(CalendarEntity $calendar, array $options): ?Response
+    protected function displayAccessCell(?ConcreteConversation $conversation, array $options): ?Response
     {
         $pk = Key::getByID($options['pkID']);
-        $pk->setPermissionObject($calendar);
+        $pk->setPermissionObject($conversation);
         $this->set('pk', $pk);
         $this->set('pa', Access::getByID($options['paID'], $pk));
         $this->setViewPath('/backend/permissions/labels');
 
         return null;
+    }
+
+    protected function savePermissionAssignments(?ConcreteConversation $conversation, array $options): ?Response
+    {
+        $permissions = Key::getList('conversations');
+        foreach ($permissions as $pk) {
+            $pk->setPermissionObject($conversation);
+            $pk->clearPermissionAssignment();
+            $paID = (int) ($options['pkID'][$pk->getPermissionKeyID()] ?? 0);
+            if ($paID > 0) {
+                $pa = Access::getByID($paID, $pk);
+                if ($pa !== null) {
+                    $pk->assignPermissionAccess($pa);
+                }
+            }
+        }
+
+        return $this->app->make(ResponseFactoryInterface::class)->json(true);
     }
 }
